@@ -46,37 +46,37 @@
   },
   "homepage": "https://github.com/PeculiarVentures/xmldsigjs#readme",
   "dependencies": {
-    "asn1js": "^2.4.0",
-    "pkijs": "^2.3.0",
+    "asn1js": "^3.0.5",
+    "pkijs": "^3.0.10",
     "pvtsutils": "^1.3.2",
     "pvutils": "^1.1.3",
-    "tslib": "^2.4.0",
-    "xml-core": "^1.1.4",
+    "tslib": "^2.4.1",
+    "xml-core": "^1.1.5",
     "xpath": "^0.0.32"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.10",
-    "@babel/plugin-proposal-class-properties": "^7.16.7",
-    "@babel/plugin-proposal-object-rest-spread": "^7.16.7",
-    "@babel/preset-env": "^7.17.10",
-    "@peculiar/webcrypto": "^1.2.3",
-    "@rollup/plugin-commonjs": "^22.0.0",
-    "@rollup/plugin-node-resolve": "^13.3.0",
-    "@rollup/plugin-typescript": "^8.3.0",
-    "@types/asn1js": "^2.0.2",
-    "@types/mocha": "^9.1.0",
-    "@types/node": "^17.0.31",
-    "@xmldom/xmldom": "^0.8.2",
+    "@babel/core": "^7.20.12",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-proposal-object-rest-spread": "^7.20.7",
+    "@babel/preset-env": "^7.20.2",
+    "@peculiar/webcrypto": "^1.4.1",
+    "@rollup/plugin-commonjs": "^24.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-typescript": "^11.0.0",
+    "@types/asn1js": "^3.0.7",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^18.11.18",
+    "@xmldom/xmldom": "^0.8.6",
     "coveralls": "^3.1.1",
-    "mocha": "^10.0.0",
+    "mocha": "^10.2.0",
     "nyc": "^15.1.0",
     "rimraf": "^3.0.2",
-    "rollup": "^2.72.0",
+    "rollup": "^3.9.1",
     "rollup-plugin-babel": "^4.4.0",
     "rollup-plugin-terser": "^7.0.2",
-    "ts-node": "^10.4.0",
+    "ts-node": "^10.9.1",
     "tslint": "^6.1.3",
-    "typescript": "^4.6.4"
+    "typescript": "^4.9.4"
   },
   "nyc": {
     "extension": [
