@@ -1,4 +1,4 @@
-import {XmlSignatureObject} from "../xml_object";
+import { XmlSignatureObject } from "../xml_object";
 
 export abstract class KeyInfoClause extends XmlSignatureObject {
     public Key: CryptoKey | null;
